requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
redis==5.0.8
psycopg2-binary==2.9.11
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.0
pydantic==2.12.5
python-dotenv==1.0.0
pytest==9.0.2
httpx==0.28.1
